SNAP-947 moved logic to selector, refactored
